interview-recorder/requirements.txt
openai-whisper
sounddevice
numpy
scipy
torch
torchvision
torchaudio
pyperclip
pynput
